OAK-8580: Add distinct cloud data store loggers for stream activity
This change allows for specific monitoring of specific named loggers
that indicate where cloud data stores are streaming bytes to or
from cloud storage.  Useful for instances that are trying to
identify where upload or download streaming through the repository
is taking place for those that are attempting to replace such
activity with direct binary access in clients.


git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1866061 13f79535-47bb-0310-9956-ffa450edef68